[mob][photos] Refactor + UI tweaks on home app bar
